refactor: break frontend circular dependencies
